drf/requirements.txt
Django>5.0,<5.2
# 파이썬은 3.13 고정
# 사용가능한버전 : 3.10, 3.11, 3.12, 3.13
djangorestframework==3.15.0,<3.15.2
django-cleanup==9.0.0
python-dotenv>=1.0.0,<2.0.0
psycopg[binary]
djangorestframework-simplejwt
django-cors-headers
Pillow>=10.0.0,<11.0.0
requests>=2.31.0,<3.0.0
pgvector>=0.3.0,<0.4.0
django-storages[s3]
boto3